requirements.txt
watson_developer_cloud==2.4.1
python-twitter==3.5
